feat(message) : modify the interfaces and account for the installer of pact-message
