+++ src/PhpPact/Standalone/Installer/Model/Scripts.php
@@ -23,17 +23,25 @@ class Scripts
     private $stubService;
 
     /**
+     * Path to the PhpPact Pact Message.
+     *
+     * @var string
+     */
+    private $pactMessage;
+
+    /**
      * Path to the PhpPact Provider Verifier.
      *
      * @var string
      */
     private $providerVerifier;
 
-    public function __construct(string $mockService, string $stubService, string $providerVerifier)
+    public function __construct(string $mockService, string $stubService, string $providerVerifier, string $pactMessage)
     {
         $this->mockService      = $mockService;
         $this->stubService      = $stubService;
         $this->providerVerifier = $providerVerifier;
+        $this->pactMessage      = $pactMessage;
     }
 
     /**
@@ -71,4 +79,13 @@ class Scripts
 
         return $this;
     }
+
+    /**
+     * @return string
+     */
+    public function getPactMessage(): string
+    {
+        return $this->pactMessage;
+    }
+
 }
